ve: Logo
