+++ src/components/vectors/Logo/Logo.tsx
@@ -1,4 +1,4 @@
-import { css, keyframes } from '~/styles'
+import styles from './Logo.css'
 
 type Props = {
   animated?: boolean
@@ -12,60 +12,11 @@ type Props = {
     }
 )
 
-const drawPath = keyframes({
-  '0%': { strokeDashoffset: 300 },
-  '100%': { strokeDashoffset: 0 }
-})
-
-const animatedClass = css({
-  '@motionOK': {
-    strokeDasharray: 300,
-    strokeDashoffset: 300,
-    willChange: 'stroke-dashoffset',
-    '& [name="b"]': {
-      animation: `${drawPath} 0.8s linear forwards`,
-      animationDelay: '0.5s'
-    },
-    '& [name="ayu"]': {
-      animation: `${drawPath} 1s ease-in-out forwards`,
-      animationDelay: '0.8s'
-    },
-    '& [name="k1"]': {
-      animation: `${drawPath} 0.4s ease-in-out forwards`,
-      animationDelay: '1.8s'
-    },
-    '& [name="k2"]': {
-      animation: `${drawPath} 0.6s ease-in-out forwards`,
-      animationDelay: '2s'
-    },
-    '& [name="ur"]': {
-      animation: `${drawPath} 0.8s linear forwards`,
-      animationDelay: '2.2s'
-    },
-    '& [name="ni"]': {
-      animation: `${drawPath} 0.7s linear forwards`,
-      animationDelay: '2.6s'
-    },
-    '& [name="dot"]': {
-      animation: `${drawPath} 0.3s linear forwards`,
-      animationDelay: '3.2s'
-    },
-    '& [name="a"]': {
-      animation: `${drawPath} 1s linear forwards`,
-      animationDelay: '2.8s'
-    },
-    '& [name="decor"]': {
-      animation: `${drawPath} 1s ease-in-out forwards`,
-      animationDelay: '3.5s'
-    }
-  }
-})()
-
 const Logo = ({ animated, className, ...props }: Props) => {
   return (
     <svg
       {...props}
-      className={animated ? `${animatedClass} ${className}` : className}
+      className={animated ? `${styles.animatedLogo} ${className}` : className}
       viewBox="0 0 147 115"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
